requirements.txt
Django~=5.2
pip~=25.0.1
attrs~=25.3.0
pyOpenSSL~=25.1.0
cryptography~=45.0.5
pillow~=11.2.1
asgiref~=3.9.1
sqlparse~=0.5.3
psycopg2-binary~=2.9.10
colorama~=0.4.6
PyYAML~=6.0.2
requests~=2.32.4
setuptools~=80.9.0
packaging~=25.0
typing_extensions~=4.13.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
PyJWT==2.10.1
drf-yasg==1.21.10

